% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{check_input_file_exists}
\alias{check_input_file_exists}
\title{check_input_file_exists}
\usage{
check_input_file_exists()
}
\value{
Character. Path to user's app data directory
}
\description{
Check if the input file exists in the user_data_dir() directory
}
