% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_var.R
\name{extract_variables}
\alias{extract_variables}
\title{Indicator variable names for a given family}
\usage{
extract_variables(x, db_variables)
}
\arguments{
\item{x}{Family name to filter on.}

\item{db_variables}{Indicator metadata table (replaces the implicit global
used by the original \code{cliarapp} version).}
}
\description{
Indicator variable names for a given family
}
