% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bivariate.R
\name{mod_bivariate_ui}
\alias{mod_bivariate_ui}
\title{bivariate correlation module UI}
\usage{
mod_bivariate_ui(id, app_data)
}
\arguments{
\item{id}{a unique identifier for this module.}

\item{app_data}{Shared data list from \code{\link[=build_app_data]{build_app_data()}}.}
}
\value{
a \code{tagList} of UI elements
}
\description{
bivariate correlation module UI
}
