% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country_comparison.R
\name{mod_country_comparison_server}
\alias{mod_country_comparison_server}
\title{country comparison module server}
\usage{
mod_country_comparison_server(id, bench, app_data)
}
\arguments{
\item{id}{a unique identifier for this module.}

\item{bench}{Named list of reactives returned by \code{\link[=mod_benchmark_server]{mod_benchmark_server()}}.}

\item{app_data}{Shared data list from \code{\link[=build_app_data]{build_app_data()}}.}
}
\value{
NULL; called for its side effects.
}
\description{
country comparison module server
}
