% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country_comparison.R
\name{mod_country_comparison_ui}
\alias{mod_country_comparison_ui}
\title{country comparison module UI}
\usage{
mod_country_comparison_ui(id, app_data)
}
\arguments{
\item{id}{a unique identifier for this module.}

\item{app_data}{Shared data list from \code{\link[=build_app_data]{build_app_data()}}.}
}
\value{
a \code{tagList} of UI elements
}
\description{
Cross-Country Comparison tab (bar chart).
}
