% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_faq.R
\name{mod_faq_ui}
\alias{mod_faq_ui}
\title{FAQ module UI}
\usage{
mod_faq_ui(id)
}
\arguments{
\item{id}{a unique identifier for this module.}
}
\value{
a \code{tagList} of UI elements
}
\description{
Static content only -- no server logic.
}
