% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{mod_home_ui}
\alias{mod_home_ui}
\title{home module UI}
\usage{
mod_home_ui(id)
}
\arguments{
\item{id}{a unique identifier for this module.}
}
\value{
a \code{tagList} of UI elements
}
\description{
Landing page. Static content only -- no server logic.
}
