% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methodology.R
\name{mod_methodology_server}
\alias{mod_methodology_server}
\title{methodology module server}
\usage{
mod_methodology_server(id, app_data)
}
\arguments{
\item{id}{a unique identifier for this module.}

\item{app_data}{Shared data list from \code{\link[=build_app_data]{build_app_data()}}.}
}
\value{
NULL; called for its side effects.
}
\description{
methodology module server
}
