% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methodology.R
\name{mod_methodology_ui}
\alias{mod_methodology_ui}
\title{methodology module UI}
\usage{
mod_methodology_ui(id, app_data)
}
\arguments{
\item{id}{a unique identifier for this module.}

\item{app_data}{Shared data list from \code{\link[=build_app_data]{build_app_data()}}.}
}
\value{
a \code{tagList} of UI elements
}
\description{
methodology module UI
}
