% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_publications.R
\name{mod_publications_server}
\alias{mod_publications_server}
\title{publications module server}
\usage{
mod_publications_server(id)
}
\arguments{
\item{id}{a unique identifier for this module. To be used in the ui and server files}
}
\value{
NULL; called for its side effect of rendering the \code{publications} output.
}
\description{
publications module server
}
