% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_publications.R
\name{mod_publications_ui}
\alias{mod_publications_ui}
\title{publications module UI}
\usage{
mod_publications_ui(id, countries)
}
\arguments{
\item{id}{a unique identifier for this module. To be used in the ui and server files}

\item{countries}{Character vector of country names for the country filter
(replaces the implicit \code{countries} global used by the original
\code{cliarapp} version).}
}
\value{
a \code{tagList} of UI elements
}
\description{
Contains code used to display content in the publications tab
}
