% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_terms.R
\name{mod_terms_ui}
\alias{mod_terms_ui}
\title{terms of use module UI}
\usage{
mod_terms_ui(id)
}
\arguments{
\item{id}{a unique identifier for this module.}
}
\value{
a \code{tagList} of UI elements
}
\description{
Static content only -- no server logic.
}
