% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trends.R
\name{mod_trends_ui}
\alias{mod_trends_ui}
\title{time trends module UI}
\usage{
mod_trends_ui(id, app_data)
}
\arguments{
\item{id}{a unique identifier for this module.}

\item{app_data}{Shared data list from \code{\link[=build_app_data]{build_app_data()}}.}
}
\value{
a \code{tagList} of UI elements
}
\description{
time trends module UI
}
