% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_world_map.R
\name{mod_world_map_ui}
\alias{mod_world_map_ui}
\title{world map module UI}
\usage{
mod_world_map_ui(id, app_data)
}
\arguments{
\item{id}{a unique identifier for this module.}

\item{app_data}{Shared data list from \code{\link[=build_app_data]{build_app_data()}}.}
}
\value{
a \code{tagList} of UI elements
}
\description{
world map module UI
}
