% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{useBs4Dash}
\alias{useBs4Dash}
\title{Use 'bs4Dash' in 'shiny'}
\usage{
useBs4Dash(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
Allow to use functions from 'bs4Dash' into a classic 'shiny' app,
specifically \code{bs4ValueBox}, \code{bs4InfoBox} and \code{bs4Card}.
}
