% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{user_data_dir}
\alias{user_data_dir}
\title{Construct user app data directory path}
\usage{
user_data_dir()
}
\value{
Character. Path to user's app data directory
}
\description{
Consists of three pieces of information:
\itemize{
\item System location
\item name
}
}
