% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{x_scatter_choices}
\alias{x_scatter_choices}
\title{This function establishes x_axis variable choices}
\usage{
x_scatter_choices(yvar, db_variables, family_names)
}
\arguments{
\item{yvar}{Selected y-axis variable name to exclude from x-axis choices.}

\item{db_variables}{Indicator metadata table (replaces the implicit global
used by the original \code{cliarapp} version).}

\item{family_names}{Data frame of family variable/name pairs (replaces the
implicit global used by the original \code{cliarapp} version).}
}
\description{
This function establishes x_axis variable choices
}
